requirements.txt
kivy>=2.1.0
kivymd>=1.0.0
pillow>=9.0.0
python-docx>=0.8.11
PyPDF2>=2.4.0
pandas>=1.3.0
openpyxl>=3.0.7
moviepy>=1.0.3
pydub>=0.25.1